+++ sdk/communication/AzureCommunicationChat/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 1.0.0
 ### Breaking Changes
-- Update message parameter `messageId` renamed to `forMessage` - `update(content: String, forMessage: String)`
+- Update message parameters updated to `update(message: String, parameters: UpdateChatMessageRequest)`
 - `EventHandler` renamed to `TrouterEventHandler`
+- Removed `AddChatParticipantsRequest`, `UpdateChatThreadRequest`, `CommunicationIdentifierModel`
 
 ## 1.0.0-beta.12 (2021-06-07)
 ### Breaking Changes